+++ srcs/ft_split_cub.c
@@ -6,7 +6,7 @@
 /*   By: rfkaier <rfkaier@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/03 18:19:45 by rfkaier           #+#    #+#             */
-/*   Updated: 2022/04/08 17:06:44 by rfkaier          ###   ########.fr       */
+/*   Updated: 2022/05/23 18:24:02 by rfkaier          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,11 +40,11 @@ char	*ft_strdupmap(const char *src, int len)
 	if (dest == NULL)
 		return (0);
 	i = 0;
-	while (i < len)
+	while ((int)i < len)
 	{
 		if (src[i] == '\0')
 		{
-			while (i < len)
+			while ((int)i < len)
 			{
 				dest[i] = '*';
 				i++;
